requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
motor
beanie
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
